Moved GUI to new folder.
