+++ src/client/gui/map_drawer.py
